Convert Log's static functions into non-static
Log's static functions were utility methods used in the custom
interface that has been removed.  For a more proper Log class,
they should no longer be static. The API of the macros stays
the same.

The static setLevel(), getLevel(), and isEnabledFor() are removed.
The static logMsg() and log() are kept because qserv needs
coordinated changes before they can be removed.