Store MDC as a single dict-like attribute of LogRecord.
If LogRecord does not have `MDC` attribute already then create it as an
instance of a special `lsst.log.utils._MDC` dictionary class. If `MDC`
attribute is already there (added by record factory) then assume it is a
dictionary and update it with MDC contents. This allows adding MDC items
at Python level as well.